Move release models into releases module
