+++ src/sentry/models/releases/constants.py
@@ -1,5 +1,1 @@
-DB_VERSION_LENGTH = 250
-
-
-ERR_RELEASE_REFERENCED = "This release is referenced by active issues and cannot be removed."
-ERR_RELEASE_HEALTH_DATA = "This release has health data and cannot be removed."
+from sentry.releases.models.releases.constants import *  # noqa: F401,F403
