+++ src/sentry/models/releases/exceptions.py
@@ -1,6 +1,1 @@
-class UnsafeReleaseDeletion(Exception):
-    pass
-
-
-class ReleaseCommitError(Exception):
-    pass
+from sentry.releases.models.releases.exceptions import *  # noqa: F401,F403
