+++ src/sentry/releases/models/release_threshold/__init__.py
@@ -1,0 +1,3 @@
+from .release_threshold import ReleaseThreshold
+
+__all__ = ("ReleaseThreshold",)
