+++ src/sentry/releases/models/releases/constants.py
@@ -1,0 +1,5 @@
+DB_VERSION_LENGTH = 250
+
+
+ERR_RELEASE_REFERENCED = "This release is referenced by active issues and cannot be removed."
+ERR_RELEASE_HEALTH_DATA = "This release has health data and cannot be removed."
