+++ src/sentry/releases/models/releases/exceptions.py
@@ -1,0 +1,6 @@
+class UnsafeReleaseDeletion(Exception):
+    pass
+
+
+class ReleaseCommitError(Exception):
+    pass
